+++ packages/webblocks-cms/public/cms/index.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-$_SERVER['SCRIPT_NAME'] = '/index.php';
-$_SERVER['PHP_SELF'] = '/index.php';
-$_SERVER['SCRIPT_FILENAME'] = dirname(__DIR__).'/index.php';
-
-require dirname(__DIR__).'/index.php';