+++ packages/webblocks-cms/routes/auth.php
@@ -4,10 +4,10 @@ use Illuminate\Support\Facades\Route;
 use WebBlocks\Cms\Http\Controllers\Auth\LoginController;
 
 Route::middleware(['web', 'guest'])->group(function () {
-  Route::get('/cms/login', [LoginController::class, 'create'])->name('login');
-  Route::post('/cms/login', [LoginController::class, 'store']);
+  Route::get('/webadmin/login', [LoginController::class, 'create'])->name('login');
+  Route::post('/webadmin/login', [LoginController::class, 'store']);
 });
 
 Route::middleware(['web', 'auth'])->group(function () {
-  Route::post('/cms/logout', [LoginController::class, 'destroy'])->name('logout');
+  Route::post('/webadmin/logout', [LoginController::class, 'destroy'])->name('logout');
 });